Fixed bugs in rebase, cleanup no longer deletes messages with attachments.

+++ MonikaBot Source/src/main/java/bot/CommandHandler.java
@@ -130,12 +130,15 @@ public class CommandHandler {
 			long ourID = MainRunner.getClient().getOurUser().getLongID();
 			for (IMessage m : history.asArray()) {
 				if (m.getAuthor().getLongID() == ourID) {
-					m.delete();
-					// Courtesy to Discord to avoid RateLimitExceptions
-					try {
-						Thread.sleep(500);
-					} catch (InterruptedException e) {
-						e.printStackTrace();
+					if(m.getAttachments().isEmpty()) {
+						m.delete();
+						
+						// Courtesy to Discord to avoid RateLimitExceptions
+						try {
+							Thread.sleep(250);
+						} catch (InterruptedException e) {
+							e.printStackTrace();
+						}
 					}
 				}
 			}
